fix: Wire.requestFrom overload, beginResponse_P deprecation, Wi-Fi scan JSON escaping
- PressureController: cast VL53L0X_ADDR to uint8_t so Wire.requestFrom resolves
  to a single overload (removes the ambiguous-call warning).
- WebConfigurator::handleRoot: use beginResponse() instead of the deprecated
  beginResponse_P() (PROGMEM handled by the regular overload in the maintained fork).
- WifiMidiHandler: JSON-escape SSIDs in the Wi-Fi scan response. A neighbouring
  SSID containing " or \ (attacker-controllable) previously broke the JSON payload.

+++ Servo_flute_ESP32/PressureController.cpp
@@ -223,7 +223,7 @@ bool PressureController::serviceTofMeasurement() {
     Wire.beginTransmission(VL53L0X_ADDR);
     Wire.write(0x13); // RESULT_INTERRUPT_STATUS
     Wire.endTransmission(false);
-    Wire.requestFrom(VL53L0X_ADDR, (uint8_t)1);
+    Wire.requestFrom((uint8_t)VL53L0X_ADDR, (uint8_t)1);
     if (Wire.available()) st = Wire.read();
     ready = (st & 0x07) != 0;
   }
